--- src/components/common/AuthProvider/PublicRoute/index.tsx
@@ -1,8 +1,0 @@
-import { Navigate, Outlet } from 'react-router-dom';
-
-const PublicRoute = () => {
-  const isAuthenticated = localStorage.getItem('token');
-  return isAuthenticated ? <Navigate to="/" replace /> : <Outlet />;
-};
-
-export default PublicRoute;
