--- src/constants/queryKey.ts
@@ -1,2 +1,0 @@
-export const useTodosKey = 'useTodosKey';
-export const useTodoKey = 'useTodoKey';
