+++ src/features/auth/utils/authValidateCheck.ts
