+++ src/features/todos/constants/index.ts
@@ -1,0 +1,4 @@
+export const QUERY_KEYS = {
+  TODO_LIST: 'TODO_LIST',
+  TODO_DETAIL: 'TODO_DETAIL',
+};
