+++ src/pages/Auth/Login/index.tsx
@@ -1,9 +1,9 @@
 import { useNavigate } from 'react-router-dom';
 
-import { login } from '@/apis/auth';
-import Form from '@/components/Form';
-import useAuthForm from '@/hooks/useAuthForm';
-import { authValidateCheck } from '@/utils/authValidateCheck';
+import { login } from '@/features/auth/api';
+import Form from '@/shared/Form';
+import useAuthForm from '@/features/auth/hooks/useAuthForm';
+import { authValidateCheck } from '@/features/auth/utils/authValidateCheck';
 
 const LoginPage = () => {
   const { data, error, handleChangeValue, handleSubmit } = useAuthForm();
@@ -18,9 +18,8 @@ const LoginPage = () => {
 
       navigate('/');
     } catch (error) {
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
       const err = error as {
-        response?: { data?: { details?: string }; status?: number };
+        response?: { data?: { details?: string } };
       };
       if (err.response?.data?.details) {
         alert(err.response?.data?.details);
